refactor: split rewriter tests by rewriter type
The single 1200-line test file made it hard to find tests for a
specific rewriter. Split into per-rewriter modules under
`tests/rewriter/`:

- katakana.rs (3 tests)
- run_rewriters.rs (3 tests, orchestrator-level)
- numeric.rs (11 tests, pure-number + counter mode)
- hiragana_variant.rs (4 tests)
- partial_hiragana.rs (6 tests)
- kanji_variant.rs (9 tests)

No behaviour change — same 36 tests, same assertions.
