+++ engine/crates/lex-core/src/converter/tests/rewriter/mod.rs
@@ -1,0 +1,6 @@
+mod hiragana_variant;
+mod kanji_variant;
+mod katakana;
+mod numeric;
+mod partial_hiragana;
+mod run_rewriters;
